Set up rough supervision tree.
The general idea behind this structure is that a request is received and this causes a query to be created. The query then creates a worker for each permutation of the mutation points and each worker actually queries the storage. The workers message the query when they have a total, and the query sums them up and returns that total to the original requester.

So, for example. A request comes in: {"AGGT", 1}. That's to find all of the sequences that could be reached by a single mutation from "AGGT". So the query launches a worker for each of the possible mutation points: ["$GGT", "A$GT", "AG$T", "AGG$"]. The records in the storage are keyed by these kinds of strings.